LayoutTests/accessibility/mac/listmarker-suffix.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/accessibility-helper.js"></script>
<script src="../../resources/js-test.js"></script>
</head>
<body>

<ul style="disc">
<li id="listitem1">item 1</li>
</ul>

<ol>
<li id="listitem2">item 2</li>
</ol>

<script>
let output = "This test makes sure the right suffix is returned when fetching text marker strings for list items.\n\n";

if (window.accessibilityController) {
    var list1 = accessibilityController.accessibleElementById("listitem1");
    var textRange = list1.textMarkerRangeForElement(list1);
    var bulletString = String.fromCharCode(0x2022);
    output += expect("list1.stringForTextMarkerRange(textRange)", "`${bulletString} item 1`");

    var list2 = accessibilityController.accessibleElementById("listitem2");
    textRange = list2.textMarkerRangeForElement(list2);
    output += expect("list2.stringForTextMarkerRange(textRange)", "'1. item 2'");

    debug(output);
}
</script>
</body>
</html>
